+++ src/algo/registration/rigid_registration/rigid_registration.h
@@ -41,7 +41,7 @@ private:
 public:
 	bool finished() override;
 	bool solveIteration() override;
-	int  currentIteration() override;
+	size_t currentIteration() override;
 	bool solve() override;
 public:
 	const SurfaceMesh & getSource() override;
